+++ src/main/kotlin/io/unthrottled/amii/AMII.kt
@@ -1,5 +1,9 @@
 package io.unthrottled.amii
 
+import com.intellij.openapi.util.IconLoader
+
 object AMII {
   const val PLUGIN_ID = "io.unthrottled.amii"
+  const val PLUGIN_NAME = "AMII"
+  val PLUGIN_ICON = IconLoader.getIcon("/icons/plugin-tool-window.svg")
 }