+++ src/main/kotlin/io/unthrottled/amii/PluginMaster.kt
@@ -1,12 +1,14 @@
-package io.unthrottled.amii.listeners
+package io.unthrottled.amii
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.ProjectManagerListener
+import io.unthrottled.amii.onboarding.UserOnBoarding
 import io.unthrottled.amii.services.MyProjectService
 
-internal class MyProjectManagerListener : ProjectManagerListener {
+internal class PluginMaster : ProjectManagerListener {
 
   override fun projectOpened(project: Project) {
     project.getService(MyProjectService::class.java)
+    UserOnBoarding.attemptToPerformNewUpdateActions(project)
   }
 }