+++ CHANGELOG.md
@@ -4,4 +4,6 @@
 
 ## [Unreleased]
 ### Added
+- Error reporting
+- New version update notification
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)